get ride  of device connect authorization on startup. closes #8
