+++ src/utils/query/wagmi.ts
@@ -1,5 +1,3 @@
-import { inAppWalletConnector } from '@thirdweb-dev/wagmi-adapter'
-import { createThirdwebClient, defineChain as thirdwebDefineChain } from 'thirdweb'
 import {
   createClient,
   formatTransactionRequest,
@@ -21,28 +19,6 @@ import { isInsideSafe } from '../safe'
 import { rainbowKitConnectors } from './wallets'
 
 const isLocalProvider = !!process.env.NEXT_PUBLIC_PROVIDER
-
-const thirdwebClientId =
-  process.env.NEXT_PUBLIC_THIRDWEB_CLIENT_ID || '4e8c81182c3709ee441e30d776223354'
-const unicornFactoryAddress =
-  process.env.NEXT_PUBLIC_NEXT_PUBLIC_UNICORN_FACTORY_ADDRESS ||
-  '0xD771615c873ba5a2149D5312448cE01D677Ee48A'
-
-// Create Thirdweb Client
-const client = createThirdwebClient({
-  clientId: thirdwebClientId,
-})
-
-// Create the Unicorn Wallet Connector (using Thirdweb In-App Wallet)
-// Note: The chain specified here is for the smart account functionality as per Unicorn docs.
-const unicornConnector = inAppWalletConnector({
-  client,
-  smartAccount: {
-    sponsorGas: true, // or false based on your needs / Unicorn requirements
-    chain: thirdwebDefineChain(mainnet.id),
-    factoryAddress: unicornFactoryAddress,
-  },
-})
 
 const tenderlyKey = process.env.NEXT_PUBLIC_TENDERLY_KEY || '4imxc4hQfRjxrVB2kWKvTo'
 const drpcKey = process.env.NEXT_PUBLIC_DRPC_KEY || 'AnmpasF2C0JBqeAEzxVO8aRuvzLTrWcR75hmDonbV6cR'
@@ -161,11 +137,9 @@ const chains = getChainsFromUrl().map((c) => ({
   },
 })) as unknown as readonly [SupportedChain, ...SupportedChain[]]
 
-const combinedConnectors = [unicornConnector, ...rainbowKitConnectors]
-
 const wagmiConfig_ = createConfig({
   syncConnectedChain: false,
-  connectors: combinedConnectors,
+  connectors: rainbowKitConnectors,
   ssr: true,
   multiInjectedProviderDiscovery: !isInsideSafe(),
   storage: createStorage({ storage: localStorageWithInvertMiddleware(), key: prefix }),
